requirements.txt
flet>=0.23.0
python-dotenv>=1.0.0
supabase>=2.0.0
pyinstaller>=6.0.0
